Build android with Makefile.
android-5.10-cutlefish.json: Input file included in manager 'kernel'
params, which contains common and module config files, along with the
exernal module location, script with additional functions, and
additional environment variables.

The environment variable 'BUILD_GOLDFISH_DRIVERS' included is used by
kbuild to add the goldfish drivers, but is not included in the kconf
config as it is considered an invalid config by the makefile.

Create-modules.sh contains functions to bundle the initramfs and zip
kernel headers.

android.go builds the kernel and modules.

Overall steps:
1) Build common kernel with common config file
2) Build modules with module config file
3) Install modules
4) Bundle initramfs and zip kernel headers